+++ addon/adapters/offline.js
@@ -429,6 +429,108 @@ export default DS.Adapter.extend({
   },
 
   /**
+    Performing batch update operaion.
+
+    @method batchUpdate
+    @param {DS.Store} store
+  	@param {DS.Model[]|DS.Model} models Is array of models or single model for batch update.
+    @return {Promise}
+  */
+  batchUpdate(store, models) {
+    let dexieService = this.get('dexieService');
+    let db = dexieService.dexie(this.get('dbName'), store);
+
+    let batchUpdateOperation = (db) => new Ember.RSVP.Promise((resolve, reject) => {
+      let promisses = Ember.A();
+      let allModelshash = Ember.Map.create();
+
+      models.forEach(model => {
+        let snapshot = model._createSnapshot();
+        let type = snapshot.type;
+        let fieldsToUpdate = (type.projections.get('ComputerE')) ? (type.projections.get('ComputerE')) : (type.projections.get('ComputerPartE'));
+        let modelHash = Ember.Map.create();
+
+        let hashOperation = (db) => new Ember.RSVP.Promise((resolve, reject) => {
+          db.table(type.modelName).get(snapshot.id).then((record) => {
+            if (!Ember.isNone(fieldsToUpdate) && record) {
+              if (!Ember.$.isEmptyObject(fieldsToUpdate)) {
+                modelHash = store.serializerFor(snapshot.modelName).serialize(snapshot, { includeId: true });
+                for (let attrName in modelHash) {
+                  if (modelHash.hasOwnProperty(attrName) && !fieldsToUpdate.hasOwnProperty(attrName)) {
+                    delete modelHash[attrName];
+                  }
+                }
+
+                // Merge record with hash and store it to local store only if hash contains some changes for record.
+                let needChangeRecord = false;
+                for (let attrName in modelHash) {
+                  if (modelHash.hasOwnProperty(attrName) && (!record[attrName] || record[attrName] !== modelHash[attrName])) {
+                    needChangeRecord = true;
+                    break;
+                  }
+                }
+
+                if (needChangeRecord) {
+                  modelHash = Ember.merge(record, modelHash);
+                } else {
+                  dexieService.set('queueSyncDownWorksCount', dexieService.get('queueSyncDownWorksCount') - 1);
+                }
+
+                resolve(snapshot.modelName);
+              } else {
+                dexieService.set('queueSyncDownWorksCount', dexieService.get('queueSyncDownWorksCount') - 1);
+                resolve();
+              }
+            } else {
+              modelHash = store.serializerFor(snapshot.modelName).serialize(snapshot, { includeId: true });
+              resolve(snapshot.modelName);
+            }
+          });
+        });
+        
+        let dexiePushPromis = new Ember.RSVP.Promise((resolve, reject) => {
+          dexieService.performOperation(db, hashOperation).then((modelName) => { 
+            let arrayOfHashes = this._calculateArrayOfHash(modelName, modelHash, allModelshash);
+            allModelshash.set(modelName, arrayOfHashes);
+
+            resolve();
+          })
+        });
+
+        promisses.push(dexiePushPromis);      
+      });
+    
+      Ember.RSVP.all(promisses).then(function() {
+        let numberOfRecordsToStore = 0;
+          if (allModelshash.size === 0) {
+            resolve();
+         } else {
+          let tableNames = allModelshash._keys.toArray();
+          db.transaction('rw', tableNames, () => {
+            for (let i = 0; i < tableNames.length; i++) {
+              let tableName = tableNames[i];
+              let arrayOfHashes = allModelshash.get(tableName);
+              numberOfRecordsToStore += arrayOfHashes ? arrayOfHashes.length : 0;
+              db.table(tableName).bulkPut(arrayOfHashes ? arrayOfHashes : {});
+            }
+          }).then(() => {
+            dexieService.set('queueSyncDownWorksCount', dexieService.get('queueSyncDownWorksCount') - numberOfRecordsToStore);
+            resolve();
+          }).catch((err) => {
+              Ember.warn('Some data loss while performing sync down records!',
+              false,
+              { id: 'ember-flexberry-data-debug.offline.sync-down-data-loss' });
+              dexieService.set('queueSyncDownWorksCount', dexieService.get('queueSyncDownWorksCount') - numberOfRecordsToStore);
+              reject(err);
+          });
+        }
+      });
+    });
+
+    return dexieService.performQueueOperation(db, batchUpdateOperation);
+  },
+
+  /**
     Stores hash for performing bulk operaion into map.
 
     @method _storeHashForBulkOperation
@@ -437,13 +539,30 @@ export default DS.Adapter.extend({
     @private
   */
   _storeHashForBulkOperation(modelName, hash) {
-    let arrayOfHashes = this._hashesToStore.get(modelName);
+    let hashes = this._hashesToStore;
+    let arrayOfHashes = this._calculateArrayOfHash(modelName, hash, hashes);
+    this._hashesToStore.set(modelName, arrayOfHashes);
+  },
+
+  /**
+    Calculate array of hashes from map.
+
+    @method _calculateArrayOfHash
+    @param {String} modelName
+    @param {Object} hash
+    @param {Ember.Map} hashes
+    @return {Array}
+    @private
+  */
+  _calculateArrayOfHash(modelName, hash, hashes) {
+    let arrayOfHashes = hashes.get(modelName);
     if (!arrayOfHashes) {
       arrayOfHashes = [];
     }
 
     arrayOfHashes.push(hash);
-    this._hashesToStore.set(modelName, arrayOfHashes);
+
+    return arrayOfHashes;
   },
 
   /**